Protein Mass Calculator
This is a small function (AAcounter()) which can calculate the mass of any protein sequence in string format. This was my solution to the Rosalind Challenge "Calculating Protein Mass".

It creates a dictionary of each amino acid's mass and frequency within the protein sequence. It then uses that data to return the mass of the protein.